use api table for item
